requirements.txt
discord.py==1.7.3
python-dotenv
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4